Fix Delete race and prioritize write ops over tirgger execution
Fixes a race condition whereby a Delete operation on a job which was mid
execution on the same scheduler instance, would not see that job as not
deleted from the in-memory.

Adds fix for this by tracking names of active jobs, and backing out of
an execution if it is no longer in that cache (deleted).

Using RWMutex, priorities Delete/Write operations over job trigger
execution when on local node.

Signed-off-by: joshvanl <me@joshvanl.dev>