API: Adds DeletePrefixes
Adds a new DeletePrefixes API which deletes all jobs (as well as
counters) whose names match the given prefixes. Note the given strings
to this API are prefixes of the _job names_, not the underlying ETCD
keys. Prefixes will be serialized into the actual ETCD keys.

This is useful for Dapr to add a feature whereby jobs (of all of types)
in a particular namespace (some job prefix) are deleted when a
Kubernetes Namespace is deleted.

Moves `/internal/tests` to `/tests` to expose cron ETCD helper funcs for
library consumers. Expands package with a `cron` sub-package to introduce
single and clustered cron server setup helpers.

Exposes the `CounterGarbageCollectionInterval` time duration pointer
option to enable testing time interval triggered counter garbage
collection. This is set to a low value during some tests, which has a
side effect of soft testing this doesn't prematurely delete counters.

Slight refactor of `api_test.go` to reside in a different package to
prevent a circular import from `/tests`.

Adds `/fake` for exposing a fake cron that library consumers can use for
testing.

Signed-off-by: joshvanl <me@joshvanl.dev>