Trigger Response Undeliverable
Adds a public API `TriggerResponse` proto message which is now returned
from the Trigger function call. This message contains a Result enum
which encompasses signalling the existing SUCCESS and FAILURE job
trigger results, as well as a new UNDELIVERABLE result.

It can be the case that a job hasn't necessarily failed, but it actually
just can't be delivered to the target at this time. To prevent the Job
going into a failure loop, receiving the `UNDELIVERABLE` result will put
the job into a new "staging" queue. This puts the job into a holding
pattern until it can be delivered.

The consumer can use the new `DeliverablePrefixes` call to register the
set of prefixes whose matching job names can be delivered. The returned
`CancelFunc` is used to de-register this prefix. Prefixes are registered
as a pool in that multiple identical prefixes can be registered and
which will remain active until all are de-registered via cancellation.

Signed-off-by: joshvanl <me@joshvanl.dev>