API engine retry (#63)
Adds a engine API retry which will continue to retry an API call if the
API currently reports that the API is closed. This is useful for when
leadership changes are in progress (typical during boot where leaders
are actively being shuffled). The retry will continue to retry an active
API call every 300ms when the API reports as closed (because leadership
has caused the API to be closed). The retry will continue to cancel
calls if the caller has cancelled the context, or if the cron is
shutting down.

Signed-off-by: joshvanl <me@joshvanl.dev>