Fix upsert of Jobs on cluster
An upsert of a Job will produce a Job whose partition will likely put
the job on another replica when running as a cluster. Previously, if a
job was not partitioned for that replica, the informer would always drop
this update event, meaning that the job would not be deleted from the
queue on the previously active replica.

Special case put events to always delete the job on the current replica
if the job it put, and the partition does not match.

Signed-off-by: joshvanl <me@joshvanl.dev>