Queue: control loop
Change refactors the main queue to not longer use semaphore locks, in
favour of using a single control loop. The control loop processes a set
of typed events sequentially, acording to the fifo enqueue order. The
control loop is responsible for processing the events and updating the
state of the triggering queue, CRUD informer events, and staging queue.

This change ensures no deadlocks occur when processing events.

Due changing the queue to no longer use inner-outer semaphore locks,
in favour of a single sequential control loop, we have performance
degradation of around ~15%. This should be able to be solved with a
follow-up change to include an inner control loop for each job.

The control loop works by processes a series of typed events. These
typed events are serialized as a oneof protobuf messgae in a new
internal `queue` proto package:

```proto
// ControlEvent is a queue event which should be executed in turn of the queue
// control loop. Used for all job events.
message ControlEvent {
  // action is the typed action which should be executed by the queue control
  // loop.
  oneof action {
    Informed informed = 1;
    ExecuteRequest execute_request = 2;
    ExecuteResponse execute_response = 3;
    DeliverablePrefixes deliverable_prefixes = 4;
    UndeliverablePrefixes undeliverable_prefixes = 5;
    Close close = 6;
  }
}
```

Consumers of the queue which makes actions against it, send these typed
messages to the queue loop.

The graveyard and collector have been removed, in favour of using
multi-operation transactions against Etcd. This ensures better
atomicity. The garbage collection option has been removed.

Signed-off-by: joshvanl <me@joshvanl.dev>