Worker Pool
Updates the trigger function to use a callback function as the result
handler, rather than the function response. This has been done to not
spawn a new go routine for every job, and instead utilize a consistent
worker pool. This significantly reduces the number of go routines in
high-throughput scenarios, improving performance and reducing
resource consumption.

Jobs are routed to a set of worker loops based on a simple hash of the
job name. By default, 64 workers are created, but this can be adjusted.

Signed-off-by: joshvanl <me@joshvanl.dev>